# api/openapi.yaml
openapi: 3.0.1
info:
  title: volunteer-portal-api
  version: 1.44.0
  contact:
    email: dev@russian.rs
    url: https://russian.rs
servers:
  - url: https://portal-test.russian.rs/api
    description: Test Environment
  - url: https://portal.russian.rs/api
    description: Prod Environment
paths:
...
  /report:
    post:
      operationId: createReport
      tags:
        - report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDto"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDto"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      operationId: updateReport
      tags:
        - report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportDto"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDto"
  /report/{id}:
    get:
      operationId: getReport
      tags:
        - report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDto"
        "404":
          $ref: "#/components/responses/ServerError"
    delete:
      operationId: deleteReport
      tags:
        - report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
        "404":
          $ref: "#/components/responses/ServerError"
  /report/{id}/note:
    post:
      operationId: addNote
      tags:
        - report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteDto"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDto"
  /report/{id}/status:
    post:
      operationId: changeStatus
      tags:
        - report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeReportStatusRequest"
      responses:
        "200":
          description: Success
  /reports:
    post:
      operationId: getReports
      tags:
        - report
      parameters:
        - name: pageRequest
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/PageRequest"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportFilter"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportPageResponse"
...
  /announcements:
    get:
      operationId: getAnnouncements
      tags:
        - announcements
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnnouncementDto"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      operationId: createAnnouncement
      tags:
        - announcements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementCreateRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementDto"
        "500":
          $ref: "#/components/responses/ServerError"
  /announcements/unread-count:
    get:
      operationId: getUnreadAnnouncementsCount
      tags:
        - announcements
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnreadAnnouncementsCountDto"
        "500":
          $ref: "#/components/responses/ServerError"
  /announcements/{id}/read:
    post:
      operationId: markAnnouncementRead
      tags:
        - announcements
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  responses:
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PageRequest:
      type: object
      description: Paging object
      properties:
        pageNumber:
          type: integer
          format: int32
          default: 0
          nullable: false
        pageSize:
          type: integer
          format: int32
          default: 10
          nullable: false
        sort:
          type: array
          default: [ ]
          items:
            type: string
    PageResponse:
      type: object
      nullable: false
      description: Page info
      required:
        - page
        - size
        - totalPages
        - totalElements
      properties:
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      description: Общий объект для сообщений об ошибках
      required:
        - message
      properties:
        message:
          type: string
          description: Описание ошибки
          nullable: false
...
    FileInfoDto:
      type: object
      nullable: true
      required:
        - id
        - name
        - link
        - size
      properties:
        id:
          type: string
          nullable: false
        name:
          type: string
          nullable: false
        link:
          type: string
          nullable: false
        contentLength:
          type: integer
          format: int64
          nullable: false
          description: Size (bytes)
    ReportFilter:
      type: object
      properties:
        login:
          type: string
          nullable: true
        dateFrom:
          type: string
          format: date
          nullable: true
        dateTo:
          type: string
          format: date
          nullable: true
        status:
          type: string
          nullable: true
        program:
          type: string
          nullable: true
          description: Показывать только отчеты по указанной программе. Фильтр работает по программе, указанной в самом отчете (на момент создания отчета), а не по текущей программе волонтера. Пустая строка значит показать отчеты без заполненной программы.
        project:
          type: string
          nullable: true
          description: Показывать только отчеты по указанному проекту. Фильтр работает по проекту, указанному в самом отчете (на момент создания отчета), а не по текущему проекту волонтера. Пустая строка значит показать отчеты без заполненного проекта.
    ReportDto:
      type: object
      nullable: false
      description: Report information
      required:
        - id
        - tasks
      properties:
        id:
          type: string
          format: uuid
        createTime:
          type: string
          format: date-time
          description: Время создания
        user:
          type: string
          description: Логин пользователя
        status:
          type: string
          description: Статус отчета
        week:
          type: number
          description: Номер отчетной недели
        program:
          type: string
          description: Код программы
          example: IT
        project:
          type: string
          description: Код проекта
          example: PORTAL
        isAuto:
          type: boolean
          nullable: false
          description: Автоматически сгенерированный отчет
          example: false
        tasks:
          type: array
          default: [ ]
          description: Список задач в отчете
          items:
            $ref: "#/components/schemas/TaskDto"
        notes:
          type: array
          default: [ ]
          description: Список заметок от модератора
          items:
            $ref: "#/components/schemas/NoteDto"
        moderator:
          type: string
          nullable: true
          description: Логин модератора
    TaskDto:
      type: object
      description: Task information
      required:
        - id
        - date
        - name
        - description
        - timeSpent
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          format: date
          description: Когда задача была выполнена
        name:
          type: string
        description:
          type: string
        nameSr:
          type: string
          nullable: true
        descriptionSr:
          type: string
          nullable: true
        customer:
          type: string
          description: Login of customer
          nullable: true
        timeSpent:
          type: integer
          description: Time spent in minutes
        result:
          type: string
          nullable: true
        files:
          type: array
          default: [ ]
          items:
            $ref: "#/components/schemas/FileInfoDto"
    ReportPageResponse:
      type: object
      nullable: false
      required:
        - page
        - content
      properties:
        page:
          $ref: "#/components/schemas/PageResponse"
        content:
          type: array
          default: [ ]
          nullable: false
          items:
            $ref: "#/components/schemas/ReportDto"
...
    NoteDto:
      type: object
      nullable: false
      required:
        - id
        - text
      properties:
        id:
          type: string
          format: uuid
        createTime:
          type: string
          format: date-time
        createdBy:
          type: string
          description: Логин пользователя, который создал заметку
          maxLength: 100
        text:
          type: string
          minLength: 1
          maxLength: 5000
    ChangeReportStatusRequest:
      type: object
      nullable: false
      required:
        - status
      properties:
        status:
          type: string
          nullable: false
        note:
          type: string
        gratitude:
          type: boolean
          description: Award curator gratitude (+points) to the volunteer when accepting the report
        managerGratitude:
          type: boolean
          description: Award manager gratitude (+points) — admins / main volunteers only
...
    AnnouncementDto:
      type: object
      required:
        - id
        - title
        - body
        - createTime
        - audience
        - read
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        body:
          type: string
        createTime:
          type: string
          format: date-time
        createdBy:
          type: string
          nullable: true
        audience:
          $ref: "#/components/schemas/AnnouncementAudience"
        programCode:
          type: string
          nullable: true
        read:
          type: boolean
    AnnouncementCreateRequest:
      type: object
      required:
        - title
        - body
        - audience
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
        body:
          type: string
          minLength: 1
          maxLength: 10000
        audience:
          $ref: "#/components/schemas/AnnouncementAudience"
        programCode:
          type: string
          nullable: true
    AnnouncementAudience:
      type: string
      enum:
        - ALL
        - PROGRAM
    UnreadAnnouncementsCountDto:
      type: object
      required:
        - count
      properties:
        count:
          type: integer
          format: int32